Spit smoke tests into three files
